cipher: Remove default padding
mbedtls uses a PKCS7 padding by default but users may prefer to use
their own function.  Crypto.Util.Padding, for example, provides such
a facility.  There is also a PyPadding project on Pypi.

This patch removes the default padding.  A ValueError is now raised
for modes that make use of padding if the data does not have the
expected size.

Closes #57.